requirements.txt

pandas
numpy
dvc
dagshub
scikit-learn
mlflow
dvc-s3
python-dotenv